tidying up Table component
