+++ src/index.ts
@@ -1,6 +1,5 @@
 import VueScreener from './components/VueScreener.vue'
-import Cell from './components/table/TableCell.vue'
-import HeaderCell from './components/table/TableHeaderCell.vue'
-import ValueCell from './components/table/TableValueCell.vue'
+import Head from './components/table/TableHead.vue'
+import Data from './components/table/TableData.vue'
 
-export { VueScreener, Cell, HeaderCell, ValueCell }
+export { VueScreener, Head, Data }
